+++ lib/src/cmcd/CmcdEventExcludedKeys.ts
@@ -6,18 +6,11 @@
  * @beta
  */
 export type CmcdEventExcludedKeys =
-	| 'bs'
 	| 'bsd'
-	| 'cmsdd'
-	| 'cmsds'
 	| 'd'
 	| 'dl'
 	| 'nor'
 	| 'ot'
-	| 'rc'
 	| 'rtp'
-	| 'smrt'
-	| 'su'
-	| 'ttfbb'
-	| 'ttlb'
-	| 'url';
+	| 'sid'
+	| 'su';