+++ lib/src/cmcd/prepareCmcdData.ts
@@ -4,18 +4,16 @@ import { SfToken } from '../structuredfield/SfToken.js';
 import { CMCD_EVENT_MODE } from './CMCD_EVENT_MODE.js';
 import { CMCD_FORMATTER_MAP } from './CMCD_FORMATTER_MAP.js';
 import { CMCD_REQUEST_MODE } from './CMCD_REQUEST_MODE.js';
-import { CMCD_RESPONSE_MODE } from './CMCD_RESPONSE_MODE.js';
 import type { CmcdData } from './CmcdData.js';
 import type { CmcdEncodeOptions } from './CmcdEncodeOptions.js';
 import type { CmcdFormatterOptions } from './CmcdFormatterOptions.js';
 import type { CmcdValue } from './CmcdValue.js';
 import { isCmcdEventKey } from './isCmcdEventKey.js';
 import { isCmcdRequestKey } from './isCmcdRequestKey.js';
-import { isCmcdResponseKey } from './isCmcdResponseKey.js';
+import { isCmcdResponseReceivedKey } from './isCmcdResponseReceivedKey.js';
 import { isCmcdV1Key } from './isCmcdV1Key.js';
 
 const filterMap = {
-	[CMCD_RESPONSE_MODE]: isCmcdResponseKey,
 	[CMCD_EVENT_MODE]: isCmcdEventKey,
 	[CMCD_REQUEST_MODE]: isCmcdRequestKey,
 };
@@ -43,6 +41,10 @@ export function prepareCmcdData(obj: Record<string, any>, options: CmcdEncodeOpt
 
 	// Filter keys based on the version, reporting mode and options
 	let keys = Object.keys(obj).filter(keyFilter);
+	
+	if (obj['e'] && obj['e'] !== 'rr') {
+		keys = keys.filter(key => !isCmcdResponseReceivedKey(key));
+	}
 
 	const filter = options.filter;
 	if (typeof filter === 'function') {
@@ -50,7 +52,7 @@ export function prepareCmcdData(obj: Record<string, any>, options: CmcdEncodeOpt
 	}
 
 	// Ensure all required keys are present before sorting
-	const needsTimestamp = reportingMode === CMCD_RESPONSE_MODE || reportingMode === CMCD_EVENT_MODE;
+	const needsTimestamp = reportingMode === CMCD_EVENT_MODE;
 	if (needsTimestamp && !keys.includes('ts')) {
 		keys.push('ts');
 	}